lab1/requirements.txt
click==8.0.4
Faker==13.3.2
Flask==2.0.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.0.3
pytest-mock==3.14.0
blinker==1.9.0